+++ src/features/backup/Backup.types.ts
@@ -7,6 +7,7 @@
  */
 
 import { PartialBackupFlagsInput } from '@/lib/graphql/generated/graphql.ts';
+import { ServerSettings } from '@/features/settings/Settings.types.ts';
 
 export enum BackupFlagGroup {
     LIBRARY = 'library',
@@ -16,3 +17,31 @@ export enum BackupFlagGroup {
 export type BackupFlag = keyof PartialBackupFlagsInput;
 
 export type BackupFlagInclusionState = Record<BackupFlag, boolean>;
+
+export type AutoBackupFlag = Pick<
+    ServerSettings,
+    | 'autoBackupIncludeCategories'
+    | 'autoBackupIncludeChapters'
+    | 'autoBackupIncludeClientData'
+    | 'autoBackupIncludeHistory'
+    | 'autoBackupIncludeManga'
+    | 'autoBackupIncludeServerSettings'
+    | 'autoBackupIncludeTracking'
+>;
+
+export type AutoBackupFlagInclusionState = Record<keyof AutoBackupFlag, boolean>;
+
+export type BackupSettingsType = Pick<
+    ServerSettings,
+    | 'backupPath'
+    | 'backupTime'
+    | 'backupInterval'
+    | 'backupTTL'
+    | 'autoBackupIncludeCategories'
+    | 'autoBackupIncludeChapters'
+    | 'autoBackupIncludeClientData'
+    | 'autoBackupIncludeHistory'
+    | 'autoBackupIncludeManga'
+    | 'autoBackupIncludeServerSettings'
+    | 'autoBackupIncludeTracking'
+>;